Ticket 7: 8-delete-comment-by-id: Implementation path DELETE api/comments/:comment_id
