+++ app.js
@@ -2,7 +2,8 @@ const express = require("express");
 const db = require("./db/connection")
 const { getTopics } = require("./controllers/topics.controller")
 const apiInfo = require('../backend_project_north/endpoints.json');
-const { getArticleById, getAllArticles, getArticleCommentsById, postComments, patchArticle } = require("./controllers/articles.controller");
+const { getArticleById, getAllArticles, patchArticle } = require("./controllers/articles.controller");
+const { postComments, getArticleCommentsById, deleteComment } = require("./controllers/comments.controller");
 
 const app = express()
 
@@ -27,6 +28,8 @@ app.get("/api/articles/:article_id/comments", getArticleCommentsById);
 app.post("/api/articles/:article_id/comments",postComments)
 
 app.patch("/api/articles/:article_id",patchArticle);
+
+app.delete("/api/comments/:comment_id",deleteComment)
 
 app.use((err, req, res, next) => {
     // console.error(err);
